requirements.txt
flask
flask-wtf
python-dotenv
pymongo
flask-security
email_validator
passlib